+++ CHANGES.md
@@ -22,6 +22,15 @@ To be released.
     parameter became `long` (was `int`).  [[#1609]]
  -  The type of `InvalidBlockBytesLengthException.BytesLength` property became
     `long` (was `int`).  [[#1609]]
+ -  Methods `IsEmpty()` and `IsFull()` of `KBucket` changed to properties
+    `IsEmpty` and `IsFull` respectively.  [[#1610]]
+ -  `MessageCodec` class renamed to `TcpMessageCodec`.  [[#1610]]
+ -  `lifetime` parameter removed from `IMessageCodec.Decode()`.
+    `messageLifespan` parameter added to constructors of `NetMQMessageCodec`
+    and `TcpMessageCodec`.  [[#1610]]
+ -  Removed unused `PeerStates` property from `Swarm<T>`.  [[#1610]]
+ -  Method name `BoundPeer.QueryAppProtocolVersion()` changed to
+    `BoundPeer.QueryAppProtocolVersionNetMQ()`.  [[#1610]]
 
 ### Backward-incompatible network protocol changes
 
@@ -47,6 +56,7 @@ To be released.
 [#1606]: https://github.com/planetarium/libplanet/pull/1606
 [#1609]: https://github.com/planetarium/libplanet/pull/1609
 [Bencodex 0.4.0-dev.20211123080042]: https://www.nuget.org/packages/Bencodex/0.4.0-dev.20211123080042
+[#1610]: https://github.com/planetarium/libplanet/pull/1610
 
 
 Version 0.21.1